requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
Deprecated==1.2.13
Django==4.0.2
django-redis==5.2.0
djangorestframework==3.13.1
idna==3.3
packaging==21.3
pyparsing==3.0.7
python-dotenv==0.19.2
python-env==1.0.0
pytz==2021.3
redis==4.1.2
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
wrapt==1.13.3
